Utility Functions

The DataPlates module also defines a number of functions that are usful when working with plates and wells.

wells

Generate the the well names of an SBS / MTP plate. Works for plates of size 6, 24, 96, 384. Well names are generated columns wise (A01, B01, ...), as it is easy to get them row-wise by just re-sorting.

using DataPlates
wells(6)

6-element Vector{String}:
 "A01"
 "B01"
 "A02"
 "B02"
 "A03"
 "B03"

DataPlates.wells — Function

wells(geometry::Int), wells(rows::Int, cols::Int)

geometry: currently one of 6, 24, 96, 384 rows, cols: number of rows and columns on plate values: vector of strings of form: "A01", "B01", ... sorted by column. It is easy to re-sort to get wells row-wise: sort(wells(96))

wellname

Some software write wellnames without zero-padding: "A1". The wellname function converts to sortable form:

wellname("B3")

"B03"

DataPlates.wellname — Function

wellname(w::string)
Adds zero-padding to well names: wellname("B3" == "B03"

row, col

Returns row name and column number from a wellname:

row("B03"), col("B03")

("B", 3)